requirement.txt
numpy==1.25.0
pandas==2.0.2
psycopg2-binary==2.9.6
pydicom==2.4.1
python-dateutil==2.8.2
pytz==2023.3
six==1.16.0
typing_extensions==4.6.3
tzdata==2023.3
